Student Recruit
Student Recruit Android Application 1.0

Developed By: Prathmesh Pardhiye

About:
University Recruit is an application which will have reviews regarding universities. It has reviews related to the history and infrastructure of the university. The crowd and campus life of the university is reviewed too. Lot of university reviewing websites do not offer faculty and pedagogy reviews which is very important while selecting a course. Our application also covers up the financial aid providing by the university if it is beneficial or not. One of the main aspect to do a master’s course is to land up a job. This application provides a review of jobs and placements provided by the university and it include both on campus and full-time jobs. There is another section called as grad school finder. It will guide the student to view the list of university according to their scores. In this section the student basically must fill up a form which has course selection option, GRE and undergrad score must be entered, work experience, technical papers if published and ratings regarding statement of purpose, letter of recommendation, extracurricular and community service. Once the student has filled all the sections then the application will list the universities according to that criteria.

Environment:  This project is built using Android Studio. The UI was developed using XML and Layout style design. Backend coding was done using Java. This application uses SQLite for storing data in the database. Apk version is available.
To open apk file navigate:
StudentRecruit\app\build\outputs\apk\debug\app-debug.apk